backend/requirements.txt
asgiref==3.11.1
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0

# Data processing
pandas>=2.0.0

# Groq AI Integration
groq==0.13.1
python-dotenv==1.0.0

# Jira Integration (optional)
jira==3.5.2
